修改

+++ calc.go
@@ -16,6 +16,6 @@ func Multiply(x, y int) int {
 }
 
 // Add 返回两个整数相除的函数
-func Divide(x, y int) int {
+func Divide(x, y int) float32 {
 	return x / y
 }
